[#86] Split Print module
Resolves #86
